@@ -67,6 +67,8 @@ import mod.jbk.build.compiler.dex.DexCompiler;
 import mod.jbk.build.compiler.resource.ResourceCompiler;
 import mod.jbk.util.LogUtil;
 import mod.jbk.util.TestkeySignBridge;
+import mod.pranav.build.JarBuilder;
+import mod.pranav.build.R8Compiler;
 import proguard.Configuration;
 import proguard.ConfigurationParser;
 import proguard.ParseException;
@@ -200,6 +202,8 @@ public class ProjectBuilder {
      */
     public void createDexFilesFromClasses() throws Exception {
         FileUtil.makeDir(yq.binDirectoryPath + File.separator + "dex");
+        if (proguard.isShrinkingEnabled() && proguard.isR8Enabled()) return;
+
         if (isD8Enabled()) {
             long savedTimeMillis = System.currentTimeMillis();
             try {
@@ -216,7 +220,7 @@ public class ProjectBuilder {
                     "--verbose",
                     "--multi-dex",
                     "--output=" + yq.binDirectoryPath + File.separator + "dex",
-                    proguard.isProguardEnabled() ? yq.classesProGuardPath : yq.compiledClassesPath
+                    proguard.isShrinkingEnabled() ? yq.proguardClassesPath : yq.compiledClassesPath
             );
 
             try {
@@ -294,7 +298,7 @@ public class ProjectBuilder {
     /**
      * @return Similar to {@link ProjectBuilder#getClasspath()}, but doesn't return some local libraries' JARs if ProGuard full mode is enabled
      */
-    public String getProGuardClasspath() {
+    public String getProguardClasspath() {
         Collection<String> localLibraryJarsWithFullModeOn = new LinkedList<>();
 
         for (HashMap<String, Object> localLibrary : mll.list) {
@@ -796,7 +800,7 @@ public class ProjectBuilder {
      */
     private void proguardAddLibConfigs(List<String> args) {
         for (Jp library : builtInLibraryManager.getLibraries()) {
-            File config = BuiltInLibraries.getLibraryProGuardConfiguration(library.a());
+            File config = BuiltInLibraries.getLibraryProguardConfiguration(library.a());
             if (config.exists()) {
                 args.add("-include");
                 args.add(config.getAbsolutePath());
@@ -810,6 +814,12 @@ public class ProjectBuilder {
      * @param args List of arguments to add R.java rules to.
      */
     private void proguardAddRjavaRules(List<String> args) {
+        FileUtil.writeFile(yq.proguardAutoGeneratedExclusions, getRJavaRules());
+        args.add("-include");
+        args.add(yq.proguardAutoGeneratedExclusions);
+    }
+
+    private String getRJavaRules() {
         StringBuilder sb = new StringBuilder("# R.java rules");
         for (Jp jp : builtInLibraryManager.getLibraries()) {
             if (jp.c() && !jp.b().isEmpty()) {
@@ -830,13 +840,45 @@ public class ProjectBuilder {
         }
         sb.append("\n");
         sb.append("-keep class ").append(yq.packageName).append(".R { *; }").append('\n');
-        FileUtil.writeFile(yq.proGuardAutoGeneratedExclusions, sb.toString());
-        args.add("-include");
-        args.add(yq.proGuardAutoGeneratedExclusions);
+        return sb.toString();
+    }
+
+    public void runR8() throws IOException {
+        long savedTimeMillis = System.currentTimeMillis();
+
+        ArrayList<String> config = new ArrayList<>();
+        config.add(ProguardHandler.ANDROID_PROGUARD_RULES_PATH);
+        config.add(yq.proguardAaptRules);
+        config.add(proguard.getCustomProguardRules());
+        var rules = new ArrayList<>(Arrays.asList(getRJavaRules().split("\n")));
+        for (Jp library : builtInLibraryManager.getLibraries()) {
+            File f = BuiltInLibraries.getLibraryProguardConfiguration(library.a());
+            if (f.exists()) {
+                config.add(f.getAbsolutePath());
+            }
+        }
+        config.addAll(mll.getPgRules());
+        ArrayList<String> jars = new ArrayList<>();
+        jars.add(yq.compiledClassesPath + ".jar");
+
+        for (HashMap<String, Object> hashMap : mll.list) {
+            String obj = hashMap.get("name").toString();
+            if (hashMap.containsKey("jarPath") && proguard.libIsProguardFMEnabled(obj)) {
+                jars.add(hashMap.get("jarPath").toString());
+            }
+        }
+        try {
+            JarBuilder.INSTANCE.generateJar(new File(yq.compiledClassesPath));
+            new R8Compiler(rules, config.toArray(new String[0]), getProguardClasspath().split(":"), jars.toArray(new String[0]), settings.getMinSdkVersion(), yq).compile();
+        } catch (Exception e) {
+            throw new IOException(e);
+        }
+        LogUtil.d(TAG, "R8 took " + (System.currentTimeMillis() - savedTimeMillis) + " ms");
     }
 
     public void runProguard() throws IOException {
         long savedTimeMillis = System.currentTimeMillis();
+
         ArrayList<String> args = new ArrayList<>();
 
         /* Include global ProGuard rules */
@@ -845,7 +887,7 @@ public class ProjectBuilder {
 
         /* Include ProGuard rules generated by AAPT2 */
         args.add("-include");
-        args.add(yq.aaptProGuardRules);
+        args.add(yq.proguardAaptRules);
 
         /* Include custom ProGuard rules */
         args.add("-include");
@@ -855,9 +897,9 @@ public class ProjectBuilder {
         proguardAddRjavaRules(args);
 
         /* Include local libraries' ProGuard rules */
-        for (String localLibraryProGuardRule : mll.getPgRules()) {
+        for (String rule : mll.getPgRules()) {
             args.add("-include");
-            args.add(localLibraryProGuardRule);
+            args.add(rule);
         }
 
         /* Include compiled Java classes (?) IT SAYS -in*jar*s, so why include .class es? */
@@ -872,16 +914,16 @@ public class ProjectBuilder {
             }
         }
         args.add("-libraryjars");
-        args.add(getProGuardClasspath());
+        args.add(getProguardClasspath());
         args.add("-outjars");
-        args.add(yq.classesProGuardPath);
+        args.add(yq.proguardClassesPath);
         if (proguard.isDebugFilesEnabled()) {
             args.add("-printseeds");
-            args.add(yq.proGuardSeedsPath);
+            args.add(yq.proguardSeedsPath);
             args.add("-printusage");
-            args.add(yq.proGuardUsagePath);
+            args.add(yq.proguardUsagePath);
             args.add("-printmapping");
-            args.add(yq.proGuardMappingPath);
+            args.add(yq.proguardMappingPath);
         }
         LogUtil.d(TAG, "About to run ProGuard with these arguments: " + args);
 
