@@ -81,12 +81,6 @@ public class yq {
     public final Zo O;
 
     /**
-     * Path of file with ProGuard rules automatically generated by aapt2,
-     * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/aapt_rules.pro
-     */
-    public final String aaptProGuardRules;
-
-    /**
      * Example content: /storage/emulated/0/.sketchware/mysc/605/bin/InternalDemo.apk.unsigned.aligned
      */
     public final String unsignedAlignedApkPath;
@@ -97,12 +91,6 @@ public class yq {
      * Example content: /storage/emulated/0/.sketchware/mysc/605/
      */
     public final String projectMyscPath;
-
-    /**
-     * ProGuarded classes.jar file's path of current project,
-     * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/classes_proguard.jar
-     */
-    public final String classesProGuardPath;
 
     public final String projectName;
 
@@ -132,22 +120,40 @@ public class yq {
     public final String packageNameAsFolders;
 
     /**
+     * Path of file with ProGuard rules automatically generated by aapt2,
+     * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/aapt_rules.pro
+     */
+    public final String proguardAaptRules;
+
+    /**
+     * Path of ProGuard rules auto-generated to exclude built-in libraries and non-Full Mode Local libraries,
+     * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/rules_generated.pro
+     */
+    public final String proguardAutoGeneratedExclusions;
+
+    /**
+     * ProGuarded classes.jar file's path of current project,
+     * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/classes_proguard.jar
+     */
+    public final String proguardClassesPath;
+
+    /**
      * Path of file containing the ProGuard mapping,
      * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/mapping.txt
      */
-    public final String proGuardMappingPath;
+    public final String proguardMappingPath;
 
     /**
      * Path of file containing ProGuard seeds,
      * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/seeds.txt
      */
-    public final String proGuardSeedsPath;
+    public final String proguardSeedsPath;
 
     /**
      * Path of file containing ProGuard usage,
      * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/usage.txt
      */
-    public final String proGuardUsagePath;
+    public final String proguardUsagePath;
 
     /**
      * Current project's ProjectSettings object
@@ -158,12 +164,6 @@ public class yq {
      * Example content: /storage/emulated/0/.sketchware/mysc/605/app/src/main/AndroidManifest.xml
      */
     public final String androidManifestPath;
-
-    /**
-     * Path of ProGuard rules auto-generated to exclude built-in libraries and non-Full Mode Local libraries,
-     * e.g. /storage/emulated/0/.sketchware/mysc/605/bin/rules_generated.pro
-     */
-    public final String proGuardAutoGeneratedExclusions;
 
     /**
      * Example content: /storage/emulated/0/.sketchware/mysc/605/app/src/main
@@ -232,12 +232,12 @@ public class yq {
         packageNameAsFolders = packageName.replaceAll("\\.", File.separator);
         binDirectoryPath = projectMyscPath + "bin";
         compiledClassesPath = binDirectoryPath + File.separator + "classes";
-        classesProGuardPath = binDirectoryPath + File.separator + "classes_proguard.jar";
-        aaptProGuardRules = binDirectoryPath + File.separator + "aapt_rules.pro";
-        proGuardSeedsPath = binDirectoryPath + File.separator + "seeds.txt";
-        proGuardUsagePath = binDirectoryPath + File.separator + "usage.txt";
-        proGuardMappingPath = binDirectoryPath + File.separator + "mapping.txt";
-        proGuardAutoGeneratedExclusions = binDirectoryPath + File.separator + "rules_generated.pro";
+        proguardClassesPath = binDirectoryPath + File.separator + "classes_proguard.jar";
+        proguardAaptRules = binDirectoryPath + File.separator + "aapt_rules.pro";
+        proguardSeedsPath = binDirectoryPath + File.separator + "seeds.txt";
+        proguardUsagePath = binDirectoryPath + File.separator + "usage.txt";
+        proguardMappingPath = binDirectoryPath + File.separator + "mapping.txt";
+        proguardAutoGeneratedExclusions = binDirectoryPath + File.separator + "rules_generated.pro";
         rJavaDirectoryPath = projectMyscPath + "gen";
         generatedFilesPath = projectMyscPath + "app" + File.separator + "src" + File.separator + "main";
         javaFilesPath = generatedFilesPath + File.separator + "java";
