@@ -1,7 +1,6 @@
 package com.besome.sketch.common;
 
 import android.app.Activity;
-import android.content.Context;
 import android.content.Intent;
 import android.content.res.Configuration;
 import android.graphics.Color;
@@ -9,6 +8,8 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.util.Pair;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
@@ -17,17 +18,28 @@ import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
+import androidx.activity.OnBackPressedCallback;
 import androidx.annotation.NonNull;
+import androidx.appcompat.widget.SearchView;
 import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.ListAdapter;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.besome.sketch.lib.base.BaseAppCompatActivity;
+import com.bumptech.glide.Glide;
 import com.google.android.material.textfield.TextInputLayout;
 import com.sketchware.remod.R;
 
 import java.io.File;
+import java.io.IOException;
+import java.lang.ref.WeakReference;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.stream.Stream;
 
 import a.a.a.KB;
 import a.a.a.MA;
@@ -40,6 +52,22 @@ import a.a.a.wq;
 import a.a.a.xB;
 
 public class ImportIconActivity extends BaseAppCompatActivity implements View.OnClickListener {
+    private static final int ICON_COLOR_BLACK = 0;
+    private static final int ICON_COLOR_GREY = 1;
+    private static final int ICON_COLOR_WHITE = 2;
+
+    private final OnBackPressedCallback searchViewCloser = new OnBackPressedCallback(true) {
+        @Override
+        public void handleOnBackPressed() {
+            setEnabled(false);
+            if (search.isActionViewExpanded()) {
+                search.collapseActionView();
+                searchView.setQuery("", true);
+            } else {
+                getOnBackPressedDispatcher().onBackPressed();
+            }
+        }
+    };
 
     private RecyclerView iconsList;
     private Button showBlackIcons;
@@ -47,6 +75,8 @@ public class ImportIconActivity extends BaseAppCompatActivity implements View.On
     private Button showWhiteIcons;
     private EditText iconName;
     private WB iconNameValidator;
+    private MenuItem search;
+    private SearchView searchView;
     private IconAdapter adapter = null;
     /**
      * Current icons' color, where 0 stands for black, 1 for grey, and 2 for white.
@@ -75,19 +105,19 @@ public class ImportIconActivity extends BaseAppCompatActivity implements View.On
                 if (iconNameValidator.b() && adapter.selectedIconPosition >= 0) {
                     Intent intent = new Intent();
                     intent.putExtra("iconName", iconName.getText().toString());
-                    intent.putExtra("iconPath", icons.get(adapter.selectedIconPosition).second);
+                    intent.putExtra("iconPath", adapter.getCurrentList().get(adapter.selectedIconPosition).second);
                     setResult(Activity.RESULT_OK, intent);
                     finish();
                 }
             } else if (id == R.id.btn_black) {
-                setIconColor(0);
+                setIconColor(ICON_COLOR_BLACK);
             } else if (id == R.id.btn_cancel) {
                 setResult(Activity.RESULT_CANCELED);
                 finish();
             } else if (id == R.id.btn_grey) {
-                setIconColor(1);
+                setIconColor(ICON_COLOR_GREY);
             } else if (id == R.id.btn_white) {
-                setIconColor(2);
+                setIconColor(ICON_COLOR_WHITE);
             }
         }
     }
@@ -95,7 +125,9 @@ public class ImportIconActivity extends BaseAppCompatActivity implements View.On
     @Override
     public void onConfigurationChanged(@NonNull Configuration newConfig) {
         super.onConfigurationChanged(newConfig);
-        ((GridLayoutManager) iconsList.getLayoutManager()).setSpanCount(getGridLayoutColumnCount());
+        if (iconsList.getLayoutManager() instanceof GridLayoutManager manager) {
+            manager.setSpanCount(getGridLayoutColumnCount());
+        }
         iconsList.requestLayout();
     }
 
@@ -143,58 +175,115 @@ public class ImportIconActivity extends BaseAppCompatActivity implements View.On
         iconNameValidator = new WB(getApplicationContext(), findViewById(R.id.ti_input), uq.b, alreadyAddedImageNames);
         iconName.setPrivateImeOptions("defaultInputmode=english;");
         k();
-        new Handler().postDelayed(() -> new InitialIconLoader(getApplicationContext()).execute(), 300L);
+        new Handler().postDelayed(() -> new InitialIconLoader(this).execute(), 300L);
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_import_icon, menu);
+        search = menu.findItem(R.id.menu_find);
+        searchView = (SearchView) search.getActionView();
+        searchView.setQueryHint("Search");
+        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                return false;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                filterIcons(newText);
+                return true;
+            }
+        });
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        if (item.getItemId() == R.id.menu_find) {
+            searchViewCloser.setEnabled(true);
+            getOnBackPressedDispatcher().addCallback(this, searchViewCloser);
+            return true;
+        }
+        return super.onOptionsItemSelected(item);
     }
 
     private void setIconName(int iconPosition) {
-        iconName.setText(icons.get(iconPosition).first);
+        iconName.setText(adapter.getCurrentList().get(iconPosition).first);
     }
 
     private void setIconColor(int colorType) {
         if (iconType != colorType) {
             iconType = colorType;
-            if (colorType == 0) {
+            if (colorType == ICON_COLOR_BLACK) {
                 showBlackIcons.setBackgroundColor(0xff33b8f5);
                 showGreyIcons.setBackgroundColor(0xffe5e5e5);
                 showWhiteIcons.setBackgroundColor(0xffe5e5e5);
-            } else if (colorType == 1) {
+            } else if (colorType == ICON_COLOR_GREY) {
                 showBlackIcons.setBackgroundColor(0xffe5e5e5);
                 showGreyIcons.setBackgroundColor(0xff33b8f5);
                 showWhiteIcons.setBackgroundColor(0xffe5e5e5);
-            } else if (colorType == 2) {
+            } else if (colorType == ICON_COLOR_WHITE) {
                 showBlackIcons.setBackgroundColor(0xffe5e5e5);
                 showGreyIcons.setBackgroundColor(0xffe5e5e5);
                 showWhiteIcons.setBackgroundColor(0xff33b8f5);
             }
-            new IconColorChangedIconLoader(getApplicationContext()).execute();
+            new IconColorChangedIconLoader(this).execute();
         }
     }
 
     private void listIcons() {
         icons = new ArrayList<>();
-        String color = "black";
-        if (iconType != 0) {
-            if (iconType == 1) {
-                color = "grey";
-            } else if (iconType == 2) {
-                color = "white";
-            }
-        }
+        String color = switch (iconType) {
+            case ICON_COLOR_BLACK -> "black";
+            case ICON_COLOR_GREY -> "grey";
+            case ICON_COLOR_WHITE -> "white";
+            default -> "black";
+        };
         String iconFolderName = "icon_" + color;
-        for (String iconName : new File(wq.getExtractedIconPackStoreLocation() + File.separator + iconFolderName).list()) {
-            icons.add(new Pair<>(
-                    iconName.substring(0, iconName.indexOf("_" + color)) + "_" + color,
-                    wq.getExtractedIconPackStoreLocation() + File.separator + iconFolderName + File.separator + iconName
-            ));
+        String iconPackStoreLocation = wq.getExtractedIconPackStoreLocation();
+        try (Stream<Path> iconFiles = Files.list(Paths.get(iconPackStoreLocation, iconFolderName))) {
+            iconFiles.map(Path::getFileName)
+                    .map(Path::toString)
+                    .forEach(iconName -> icons.add(new Pair<>(
+                            iconName.substring(0, iconName.indexOf("_" + color)) + "_" + color,
+                            Paths.get(iconPackStoreLocation, iconFolderName, iconName).toString()
+                    )));
+        } catch (IOException e) {
+            e.printStackTrace();
         }
     }
 
-    private class IconAdapter extends RecyclerView.Adapter<IconAdapter.ViewHolder> {
+    private void filterIcons(String query) {
+        var filteredIcons = new ArrayList<Pair<String, String>>(icons.size());
+        for (Pair<String, String> icon : icons) {
+            if (icon.first.toLowerCase().contains(query.toLowerCase())) {
+                filteredIcons.add(icon);
+            }
+        }
+        adapter.submitList(filteredIcons);
+    }
 
+    private class IconAdapter extends ListAdapter<Pair<String, String>, IconAdapter.ViewHolder> {
+        private static final DiffUtil.ItemCallback<Pair<String, String>> DIFF_CALLBACK = new DiffUtil.ItemCallback<>() {
+            @Override
+            public boolean areItemsTheSame(@NonNull Pair<String, String> oldItem, @NonNull Pair<String, String> newItem) {
+                return oldItem.first.equals(newItem.first);
+            }
+
+            @Override
+            public boolean areContentsTheSame(@NonNull Pair<String, String> oldItem, @NonNull Pair<String, String> newItem) {
+                return true;
+            }
+        };
         private int selectedIconPosition = -1;
 
-        private class ViewHolder extends RecyclerView.ViewHolder {
+        protected IconAdapter() {
+            super(DIFF_CALLBACK);
+        }
 
+        private class ViewHolder extends RecyclerView.ViewHolder {
             public final RelativeLayout background;
             public final TextView name;
             public final ImageView icon;
@@ -219,7 +308,7 @@ public class ImportIconActivity extends BaseAppCompatActivity implements View.On
         @Override
         public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
             if (position != selectedIconPosition) {
-                if (iconType == 2) {
+                if (iconType == ICON_COLOR_WHITE) {
                     holder.background.setBackgroundColor(0xffbdbdbd);
                 } else {
                     holder.background.setBackgroundColor(Color.WHITE);
@@ -227,12 +316,10 @@ public class ImportIconActivity extends BaseAppCompatActivity implements View.On
             } else {
                 holder.background.setBackgroundColor(0xffffccbc);
             }
-            holder.name.setText(icons.get(position).first);
-            try {
-                holder.icon.setImageBitmap(iB.a(icons.get(position).second, 1));
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
+            holder.name.setText(getItem(position).first);
+            Glide.with(ImportIconActivity.this)
+                    .load(getItem(position).second)
+                    .into(holder.icon);
         }
 
         @Override
@@ -240,35 +327,35 @@ public class ImportIconActivity extends BaseAppCompatActivity implements View.On
         public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
             return new ViewHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.import_icon_list_item, parent, false));
         }
-
-        @Override
-        public int getItemCount() {
-            return icons.size();
-        }
     }
 
-    private class InitialIconLoader extends MA {
-        public InitialIconLoader(Context context) {
-            super(context);
-            addTask(this);
+    private static class InitialIconLoader extends MA {
+        private final WeakReference<ImportIconActivity> activity;
+
+        public InitialIconLoader(ImportIconActivity activity) {
+            super(activity);
+            this.activity = new WeakReference<>(activity);
+            activity.addTask(this);
         }
 
         @Override
         public void a() {
-            h();
-            setIconColor(0);
+            var activity = this.activity.get();
+            activity.h();
+            activity.setIconColor(ICON_COLOR_BLACK);
         }
 
         @Override
         public void b() {
-            if (!doExtractedIconsExist()) {
-                extractIcons();
+            var activity = this.activity.get();
+            if (!activity.doExtractedIconsExist()) {
+                activity.extractIcons();
             }
         }
 
         @Override
         public void a(String str) {
-            h();
+            activity.get().h();
         }
 
         @Override
@@ -277,29 +364,41 @@ public class ImportIconActivity extends BaseAppCompatActivity implements View.On
         }
     }
 
-    private class IconColorChangedIconLoader extends MA {
-        public IconColorChangedIconLoader(Context context) {
-            super(context);
-            addTask(this);
-            k();
+    private static class IconColorChangedIconLoader extends MA {
+        private final WeakReference<ImportIconActivity> activity;
+
+        public IconColorChangedIconLoader(ImportIconActivity activity) {
+            super(activity);
+            this.activity = new WeakReference<>(activity);
+            activity.addTask(this);
+            activity.k();
         }
 
         @Override
         public void a() {
-            h();
-            iconName.setText("");
-            adapter.selectedIconPosition = -1;
-            adapter.notifyDataSetChanged();
+            var activity = this.activity.get();
+            activity.h();
+            activity.iconName.setText("");
+            int oldPosition = activity.adapter.selectedIconPosition;
+            activity.adapter.selectedIconPosition = -1;
+            activity.adapter.notifyItemChanged(oldPosition);
+            activity.adapter.submitList(activity.icons);
         }
 
         @Override
         public void b() {
-            listIcons();
+            var activity = this.activity.get();
+            activity.listIcons();
+            activity.runOnUiThread(() -> {
+                if (activity.searchView != null) {
+                    activity.searchView.setQuery("", false);
+                }
+            });
         }
 
         @Override
         public void a(String str) {
-            h();
+            activity.get().h();
         }
 
         @Override
