@@ -163,8 +163,10 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         }
     });
     public final ActivityResultLauncher<Intent> changeOpenFile = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), result -> {
-        assert result.getData() != null;
-        projectFileSelector.setXmlFileName(result.getData().getParcelableExtra("project_file"));
+        if (result.getResultCode() == Activity.RESULT_OK) {
+            assert result.getData() != null;
+            projectFileSelector.setXmlFileName(result.getData().getParcelableExtra("project_file"));
+        }
     });
 
     /**
@@ -803,7 +805,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
     /**
      * Opens {@link ManageProguardActivity}.
      */
-    void toProGuardManager() {
+    void toProguardManager() {
         launchActivity(ManageProguardActivity.class, null);
     }
 
