@@ -6,11 +6,14 @@ import android.util.AttributeSet;
 import androidx.annotation.NonNull;
 
 import com.besome.sketch.lib.base.CollapsibleLayout;
+import com.besome.sketch.lib.ui.CollapsibleButton;
 import com.sketchware.remod.R;
 
 import java.util.List;
 
-public class CollapsibleEventLayout extends CollapsibleLayout {
+import mod.hey.studios.util.Helper;
+
+public class CollapsibleEventLayout extends CollapsibleLayout<CollapsibleButton> {
     private CollapsibleButton delete;
     private CollapsibleButton addToCollection;
 
@@ -29,6 +32,21 @@ public class CollapsibleEventLayout extends CollapsibleLayout {
         addToCollection = CollapsibleButton.create(context, 2, R.drawable.ic_bookmark_red_48dp, R.string.logic_list_menu_add_to_collection);
         addToCollection.setVisibility(GONE);
         return List.of(reset, delete, addToCollection);
+    }
+
+    @Override
+    protected CharSequence getWarningMessage() {
+        return Helper.getResString(this, R.string.common_message_confirm);
+    }
+
+    @Override
+    protected CharSequence getYesLabel() {
+        return Helper.getResString(this, R.string.common_word_continue);
+    }
+
+    @Override
+    protected CharSequence getNoLabel() {
+        return Helper.getResString(this, R.string.common_word_cancel);
     }
 
     public void hideDelete() {
