@@ -1,81 +1,35 @@
 package com.besome.sketch.editor.manage.library.googlemap;
 
-import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.os.Build.VERSION;
 import android.os.Bundle;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.Switch;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.widget.Toolbar;
-import androidx.core.content.FileProvider;
-import androidx.recyclerview.widget.DefaultItemAnimator;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 import com.besome.sketch.beans.ProjectLibraryBean;
-import com.besome.sketch.editor.manage.library.ProjectComparator;
 import com.besome.sketch.lib.base.BaseAppCompatActivity;
-import com.besome.sketch.lib.ui.CircleImageView;
 import com.sketchware.remod.R;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 
 import a.a.a.GB;
 import a.a.a.aB;
 import a.a.a.bB;
 import a.a.a.iC;
-import a.a.a.lC;
 import a.a.a.mB;
-import a.a.a.wB;
-import a.a.a.wq;
-import a.a.a.yB;
 import mod.hey.studios.util.Helper;
+import mod.jbk.editor.manage.library.LibrarySettingsImporter;
 
 public class ManageGoogleMapActivity extends BaseAppCompatActivity implements View.OnClickListener {
-
     private String sc_id;
     private Switch libSwitch;
     private EditText editApiKey;
     private ProjectLibraryBean googleMapLibraryBean;
-    private ArrayList<HashMap<String, Object>> projectsList = new ArrayList<>();
-    private ProjectAdapter projectAdapter;
-
-    private void initializeProjectList() {
-        projectsList = new ArrayList<>();
-
-        for (HashMap<String, Object> projectMap : lC.a()) {
-            String projectScId = yB.c(projectMap, "sc_id");
-            if (!sc_id.equals(projectScId)) {
-                iC projectLibraryHandler = new iC(projectScId);
-                projectLibraryHandler.i();
-                if (projectLibraryHandler.e().useYn.equals("Y")) {
-                    projectMap.put("google_map", projectLibraryHandler.e().clone());
-                    projectsList.add(projectMap);
-                }
-            }
-        }
-
-        if (projectsList.size() > 0) {
-            //noinspection Java8ListSort
-            Collections.sort(projectsList, new ProjectComparator());
-        }
-
-        projectAdapter.notifyDataSetChanged();
-    }
 
     private void openDoc() {
         if (GB.h(getApplicationContext())) {
@@ -95,33 +49,6 @@ public class ManageGoogleMapActivity extends BaseAppCompatActivity implements Vi
         } else {
             bB.a(getApplicationContext(), Helper.getResString(R.string.common_message_check_network), 0).show();
         }
-    }
-
-    private void importLibrarySettings() {
-        final aB dialog = new aB(this);
-        dialog.b(Helper.getResString(R.string.design_library_title_select_project));
-        dialog.a(R.drawable.widget_google_map);
-        View rootView = wB.a(this, R.layout.manage_library_popup_project_selector);
-        RecyclerView projectRecyclerView = rootView.findViewById(R.id.list);
-        projectRecyclerView.setHasFixedSize(true);
-        projectRecyclerView.setLayoutManager(new LinearLayoutManager(this));
-        projectAdapter = new ProjectAdapter();
-        projectRecyclerView.setAdapter(projectAdapter);
-        projectRecyclerView.setItemAnimator(new DefaultItemAnimator());
-        initializeProjectList();
-        dialog.a(rootView);
-        dialog.b(Helper.getResString(R.string.common_word_select), view -> {
-            if (!mB.a()) {
-                if (projectAdapter.selectedProjectIndex >= 0) {
-                    HashMap<String, Object> projectMap = projectsList.get(projectAdapter.selectedProjectIndex);
-                    googleMapLibraryBean = (ProjectLibraryBean) projectMap.get("google_map");
-                    configure();
-                    dialog.dismiss();
-                }
-            }
-        });
-        dialog.a(Helper.getResString(R.string.common_word_cancel), Helper.getDialogDismissListener(dialog));
-        dialog.show();
     }
 
     private void downloadChromeDialog() {
@@ -160,7 +87,12 @@ public class ManageGoogleMapActivity extends BaseAppCompatActivity implements Vi
         int id = v.getId();
 
         if (id == R.id.btn_import) {
-            importLibrarySettings();
+            LibrarySettingsImporter importer = new LibrarySettingsImporter(sc_id, iC::e);
+            importer.addOnProjectSelectedListener(settings -> {
+                googleMapLibraryBean = settings;
+                configure();
+            });
+            importer.showDialog(this);
         } else if (id == R.id.btn_open_doc) {
             openDoc();
         } else if (id == R.id.layout_switch) {
@@ -233,94 +165,5 @@ public class ManageGoogleMapActivity extends BaseAppCompatActivity implements Vi
     private void configure() {
         libSwitch.setChecked("Y".equals(googleMapLibraryBean.useYn));
         editApiKey.setText(googleMapLibraryBean.data);
-    }
-
-    private class ProjectAdapter extends RecyclerView.Adapter<ProjectAdapter.ViewHolder> {
-        private int selectedProjectIndex = -1;
-
-        @Override
-        public int getItemCount() {
-            return projectsList.size();
-        }
-
-        @Override
-        public void onBindViewHolder(@NonNull ViewHolder viewHolder, int position) {
-            HashMap<String, Object> projectMap = projectsList.get(position);
-            String sc_id = yB.c(projectMap, "sc_id");
-            viewHolder.imgIcon.setImageResource(R.drawable.default_icon);
-            if (yB.a(projectMap, "custom_icon")) {
-                Uri iconUri;
-                if (VERSION.SDK_INT >= 24) {
-                    Context applicationContext =
-                            getApplicationContext();
-                    String providerPath = getPackageName() + ".provider";
-                    String iconPath = wq.e() + File.separator + sc_id;
-                    iconUri =
-                            FileProvider.getUriForFile(applicationContext, providerPath, new File(iconPath, "icon.png"));
-                } else {
-                    String iconPath = wq.e() + File.separator + sc_id;
-                    iconUri = Uri.fromFile(new File(iconPath, "icon.png"));
-                }
-
-                viewHolder.imgIcon.setImageURI(iconUri);
-            }
-
-            viewHolder.appName.setText(yB.c(projectMap, "my_app_name"));
-            viewHolder.projectName.setText(yB.c(projectMap, "my_ws_name"));
-            viewHolder.pkgName.setText(yB.c(projectMap, "my_sc_pkg_name"));
-            String version =
-                    String.format(
-                            "%s(%s)",
-                            yB.c(projectMap, "sc_ver_name"), yB.c(projectMap, "sc_ver_code"));
-            viewHolder.projectVersion.setText(version);
-            viewHolder.imgSelected.setVisibility(yB.a(projectMap, "selected") ? View.VISIBLE : View.GONE);
-        }
-
-        @Override
-        @NonNull
-        public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-            return new ViewHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.manage_library_popup_project_list_item, parent, false));
-        }
-
-        private class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
-            public final LinearLayout projectLayout;
-            public final CircleImageView imgIcon;
-            public final TextView projectName;
-            public final TextView appName;
-            public final TextView pkgName;
-            public final TextView projectVersion;
-            public final ImageView imgSelected;
-
-            public ViewHolder(View itemView) {
-                super(itemView);
-                projectLayout = itemView.findViewById(R.id.project_layout);
-                projectName = itemView.findViewById(R.id.project_name);
-                imgIcon = itemView.findViewById(R.id.img_icon);
-                appName = itemView.findViewById(R.id.app_name);
-                pkgName = itemView.findViewById(R.id.package_name);
-                projectVersion = itemView.findViewById(R.id.project_version);
-                imgSelected = itemView.findViewById(R.id.img_selected);
-                projectLayout.setOnClickListener(this);
-            }
-
-            @Override
-            public void onClick(View v) {
-                if (!mB.a() && v.getId() == R.id.project_layout) {
-                    selectedProjectIndex = getLayoutPosition();
-                    selectProject(selectedProjectIndex);
-                }
-            }
-
-            private void selectProject(int index) {
-                if (projectsList.size() > 0) {
-                    for (HashMap<String, Object> projectMap : projectsList) {
-                        projectMap.put("selected", false);
-                    }
-
-                    projectsList.get(index).put("selected", true);
-                    notifyDataSetChanged();
-                }
-            }
-        }
     }
 }
