@@ -147,7 +147,7 @@ public class ProguardHandler {
         return debugFiles;
     }
 
-    public boolean isProguardEnabled() {
+    public boolean isShrinkingEnabled() {
         boolean proguardEnabled = true;
         if (FileUtil.isExistFile(config_path)) {
             try {
@@ -175,9 +175,37 @@ public class ProguardHandler {
         FileUtil.writeFile(config_path, new Gson().toJson(config));
     }
 
+    public boolean isR8Enabled() {
+        boolean r8Enabled = true;
+        if (FileUtil.isExistFile(config_path)) {
+            try {
+                var config = new Gson().fromJson(FileUtil.readFile(config_path), Helper.TYPE_STRING_MAP);
+
+                String enabled = config.get("r8");
+                if (enabled == null) {
+                    r8Enabled = false;
+                } else {
+                    r8Enabled = enabled.equals("true");
+                }
+
+            } catch (Exception e) {
+                r8Enabled = false;
+            }
+        }
+
+        return r8Enabled;
+    }
+
+    public void setR8Enabled(boolean r8Enabled) {
+        var config = new Gson().fromJson(FileUtil.readFile(config_path), Helper.TYPE_STRING_MAP);
+        config.put("r8", String.valueOf(r8Enabled));
+
+        FileUtil.writeFile(config_path, new Gson().toJson(config));
+    }
+
     public boolean libIsProguardFMEnabled(String library) {
         boolean enabled;
-        if (isProguardEnabled() && FileUtil.isExistFile(fm_config_path)) {
+        if (isShrinkingEnabled() && FileUtil.isExistFile(fm_config_path)) {
             String configContent = FileUtil.readFile(fm_config_path);
 
             if (configContent.isEmpty()) {
@@ -207,9 +235,14 @@ public class ProguardHandler {
     }
 
     public void start(BuildProgressReceiver progressReceiver, ProjectBuilder builder) throws IOException {
-        if (isProguardEnabled()) {
-            progressReceiver.onProgress("ProGuarding classes...");
-            builder.runProguard();
+        if (isShrinkingEnabled()) {
+            if (isR8Enabled()) {
+                progressReceiver.onProgress("Running R8 on classes...");
+                builder.runR8();
+            } else {
+                progressReceiver.onProgress("ProGuarding classes...");
+                builder.runProguard();
+            }
         }
     }
 }
